refactor: di in data module (#90)
